feat: support remove component snippet
