+++ extensions/iceworks-refactor/package.nls.json
@@ -1,3 +1,4 @@
 {
-  "iceworks-refactor.component.remove.title": "Iceworks: Remove Component And the Reference"
+  "iceworks-refactor.component.dir-and-reference.remove.title": "Iceworks: Remove Component And the Reference",
+  "iceworks-refactor.component.snippet-and-reference.remove.title": "Iceworks: Remove Component Snippet And the Reference"
 }
