+++ extensions/iceworks-refactor/src/commands/index.ts
@@ -1,0 +1,2 @@
+export { default as removeCompAndRef } from './removeCompAndRef';
+export { default as removeCompSnippetAndRef } from './removeCompSnippetAndRef';
