+++ extensions/iceworks-refactor/src/extension.ts
@@ -1,12 +1,13 @@
 import { ExtensionContext, commands } from 'vscode';
 import { initExtension } from '@iceworks/common-service';
-import removeComponentAndReference from './commands/removeComponentAndReference';
+import { removeCompAndRef, removeCompSnippetAndRef } from './commands';
 
 function activate(context: ExtensionContext) {
   // auto set configuration
   initExtension(context);
 
-  commands.registerCommand('iceworks-refactor.component.remove', removeComponentAndReference);
+  commands.registerCommand('iceworks-refactor.component.dir-and-reference.remove', removeCompAndRef);
+  commands.registerTextEditorCommand('iceworks-refactor.component.snippet-and-reference.remove', removeCompSnippetAndRef);
 }
 
 exports.activate = activate;
