+++ extensions/iceworks-refactor/src/refactor/parsers/index.ts
@@ -1,11 +1,4 @@
-import findImportSpecifiers from './findImportSpecifiers';
-import findUnreferencedIdentifiers from './findUnreferencedIdentifiers';
-import removeElement from './removeElement';
-import removeDeadReferences from './removeDeadReferences';
-
-export {
-  findImportSpecifiers,
-  findUnreferencedIdentifiers,
-  removeElement,
-  removeDeadReferences,
-};
+export { default as findImportSpecifiers } from './findImportSpecifiers';
+export { default as findUnreferencedIdentifiers } from './findUnreferencedIdentifiers';
+export { default as removeElement } from './removeElement';
+export { default as removeDeadReferences } from './removeDeadReferences';
