+++ extensions/iceworks-refactor/src/refactor/utils/generate.ts
@@ -1,7 +1,11 @@
 import generate from '@babel/generator';
 
+const options = {
+  retainLines: true,
+};
+
 function generateCode(ast) {
-  return generate(ast).code;
+  return generate(ast, options).code;
 }
 
 export default generateCode;
