+++ extensions/iceworks-refactor/src/test/suite/refactor.test.ts
@@ -1,7 +1,7 @@
 import * as assert from 'assert';
 import * as fs from 'fs';
 import * as path from 'path';
-import { removeComponentCode } from '../../refactor';
+import { removeComponent } from '../../refactor';
 import prettierFormat from '../../utils/prettierFormat';
 
 const testPath = path.resolve(__dirname, '../../../', 'src/test');
@@ -12,7 +12,7 @@ suite('Refactor: remove component in page', () => {
   test('Remove component in function component', (done) => {
     const componentPath = path.join(examplesPath, 'components', 'Detail', 'index.tsx');
     const pagePath = path.join(examplesPath, 'pages', 'DetailPage', 'index.tsx');
-    const code = removeComponentCode(pagePath, componentPath, 'ts');
+    const code = removeComponent(pagePath, componentPath, 'ts');
     const expectCode = prettierFormat(fs.readFileSync(path.join(expectsPath, 'DetailPage.tsx'), { encoding: 'utf-8' }));
 
     assert.strictEqual(code, expectCode);
@@ -22,7 +22,7 @@ suite('Refactor: remove component in page', () => {
   test('Remove component in class component', (done) => {
     const componentPath = path.join(examplesPath, 'components', 'Todo', 'index.tsx');
     const pagePath = path.join(examplesPath, 'pages', 'TodoPage', 'index.tsx');
-    const code = removeComponentCode(pagePath, componentPath, 'ts');
+    const code = removeComponent(pagePath, componentPath, 'ts');
     const expectCode = prettierFormat(fs.readFileSync(path.join(expectsPath, 'TodoPage.tsx'), { encoding: 'utf-8' }));
 
     assert.strictEqual(code, expectCode);
