+++ extensions/iceworks-refactor/src/utils/prettierFormat.ts
@@ -4,7 +4,7 @@ function prettierFormat(code: string) {
   return prettier.format(code, {
     singleQuote: true,
     trailingComma: 'es5',
-    parser: 'typescript',
+    parser: 'babel',
   });
 }
